Fix download button ahead of adobe brackets EOL
Adobe will be deleting the adobe/brackets repo at some point soon, which is a problem because the download button still downloads the binaries from there. This commit  adds the version files to this repo so that the site will not break on september 1st. The binaries have been copied to scanunicco/brackets as they cannot be added to brackets-cont/brackets vie a PR. This is just a temporary fix until we upload the releases the regular brackets-cont/brackets repo.